UI inicial react

+++ FinanceControl/FinanceControl.Application/DTOs/CategoryDto.cs
@@ -4,7 +4,7 @@
     {
         public string? Id { get; set; }
         public string Name { get; set; } = string.Empty;
-        public float Limit { get; set; }
+        public decimal Limit { get; set; }
         public int Type { get; set; }  // enum convertido para int
     }
 }
