+++ FinanceControl/FinanceControl.Application/DTOs/CreateCategoryDto.cs
@@ -3,7 +3,7 @@
     public class CreateCategoryDto
     {
         public string Name { get; set; } = string.Empty;
-        public float Limit { get; set; }
+        public decimal Limit { get; set; }
         public int Type { get; set; }
     }
 }
